rbac acl user aclModule

+++ system/back-end/src/main/java/com/cy/single/blog/service/impl/SysAclModuleServiceImpl.java
@@ -8,8 +8,11 @@ import com.cy.single.blog.dao.SysAclModuleMapper;
 import com.cy.single.blog.pojo.dto.sys.aclmodule.AclModuleDto;
 import com.cy.single.blog.pojo.entity.sys.SysAclModule;
 import com.cy.single.blog.pojo.req.aclmodule.AclModuleDelReq;
+import com.cy.single.blog.pojo.req.aclmodule.AclModuleListReq;
 import com.cy.single.blog.pojo.req.aclmodule.AclModuleReq;
+import com.cy.single.blog.pojo.vo.sys.aclmodule.SysAclModuleVO;
 import com.cy.single.blog.service.SysAclModuleService;
+import com.cy.single.blog.service.SysAclService;
 import com.cy.single.blog.service.SysTreeService;
 import com.cy.single.blog.utils.dateUtil.DateUtil;
 import com.cy.single.blog.utils.keyUtil.IdWorker;
@@ -17,6 +20,7 @@ import com.cy.single.blog.utils.orgUtil.LevelUtil;
 import com.google.common.base.Preconditions;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,6 +28,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
+
+import static com.cy.single.blog.enums.ReturnCodeEnum.DATA_INFO_UNUSUAL;
+import static com.cy.single.blog.enums.ReturnCodeEnum.INFO_NOT_EXIST;
 
 /**
  * @Author: Lil-K
@@ -36,6 +43,9 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 
 	@Autowired
 	private SysAclModuleMapper aclModuleMapper;
+
+	@Autowired
+	private SysAclService aclService;
 
 	@Autowired
 	private SysTreeService treeService;
@@ -55,20 +65,27 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 		}
 
 		/**计算层级**/
-		String level = LevelUtil.calculateLevel(getLevel(req.getParentId()),req.getParentId());
+		SysAclModule parentAclModule = getParentAclModule(req.getParentSurrogateId());
+		if (Objects.isNull(parentAclModule)) {
+			return ApiResp.failure(DATA_INFO_UNUSUAL);
+		}
+		String level = LevelUtil.calculateLevel(Objects.isNull(parentAclModule) ? null : parentAclModule.getLevel(), parentAclModule.getId());
+
 		Long surrogateId = IdWorker.getSnowFlakeId(); // surrogateId
 		Date currentTime = DateUtil.localDateTimeNow();// 当前时间
 		SysAclModule aclModule = SysAclModule.builder()
 			.surrogateId(surrogateId)
 			.number("ACLM"+ surrogateId)
-			.parentId(req.getParentSurrogateId())
-			.seq(req.getSeq())
-			.level(level)
+			.parentId(parentAclModule.getSurrogateId())
+			.parentName(parentAclModule.getName())
 			.name(req.getName())
+			.level(level)
+			.seq(req.getSeq())
 			.remark(req.getRemark())
 			.createTime(currentTime)
 			.updateTime(currentTime)
 			.operator(RequestHolder.getCurrentUser().getSurrogateId())
+			.creatorId(RequestHolder.getCurrentUser().getSurrogateId())
 			.operateIp("127.0.0.1")
 			.build();
 		aclModuleMapper.insert(aclModule);
@@ -79,17 +96,17 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 	 * 检查部权限模块是否存在
 	 * @param parentId
 	 * @param aclModuleName
-	 * @param SurrogateId
+	 * @param surrogateId
 	 * @return true/false
 	 */
-	protected boolean checkAclModuleExist(Long parentId, String aclModuleName, Long SurrogateId){
+	protected boolean checkAclModuleExist(Long parentId, String aclModuleName, Long surrogateId){
 		QueryWrapper<SysAclModule> query1 = new QueryWrapper<>();
 		query1.eq("parent_id",parentId);
 		if (Objects.nonNull(aclModuleName)) {
 			query1.eq("name",aclModuleName);
 		}
-		if (Objects.nonNull(SurrogateId)) {
-			query1.eq("surrogate_id",SurrogateId);
+		if (Objects.nonNull(surrogateId)) {
+			query1.eq("surrogate_id",surrogateId);
 		}
 		Long count = aclModuleMapper.selectCount(query1);
 		if (count >= 1) {
@@ -101,16 +118,17 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 
 	/**
 	 * 获取当前权限模块所在层级的level
-	 * @param id
+	 * @param surrogateId
 	 * @return
 	 */
-	private String getLevel(Long id) {
-		SysAclModule aclModule = aclModuleMapper.selectById(id);
+	private SysAclModule getParentAclModule(Long surrogateId) {
+		QueryWrapper<SysAclModule> query = new QueryWrapper<>();
+		query.eq("surrogate_id", surrogateId);
+		SysAclModule aclModule = aclModuleMapper.selectOne(query);
 		if (Objects.isNull(aclModule)) {
 			return null;
-		}else {
-			return aclModule.getLevel();
 		}
+		return aclModule;
 	}
 
 	/**
@@ -123,8 +141,13 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 		/**
 		 * 检查权限模块名是否相同
 		 */
-		if (checkAclModuleExist(req.getParentSurrogateId(),req.getName(),req.getSurrogateId())) {
-			return ApiResp.failure("待修改的权限模块名不能重复");
+//		if (checkAclModuleExist(req.getParentSurrogateId(), req.getName(), req.getSurrogateId())) {
+//			return ApiResp.failure("待修改的权限模块名不能重复");
+//		}
+
+		SysAclModule parentAclModule = getParentAclModule(req.getParentSurrogateId());
+		if (Objects.isNull(parentAclModule)) {
+			return ApiResp.failure(DATA_INFO_UNUSUAL);
 		}
 
 		// 检查待更新的权限模块是否存在
@@ -138,9 +161,10 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 			.id(before.getId())
 			.surrogateId(before.getSurrogateId())
 			.name(req.getName())
-			.parentId(req.getParentSurrogateId())
+			.parentId(parentAclModule.getSurrogateId())
+			.parentName(parentAclModule.getName())
 			.seq(req.getSeq())
-			.level(LevelUtil.calculateLevel(getLevel(req.getParentId()), req.getParentId()))
+			.level(LevelUtil.calculateLevel(parentAclModule.getLevel(), parentAclModule.getId()))
 			.remark(req.getRemark())
 			.updateTime(DateUtil.localDateTimeNow())
 			.operator(RequestHolder.getCurrentUser().getSurrogateId())
@@ -224,7 +248,40 @@ public class SysAclModuleServiceImpl extends ServiceImpl<SysAclModuleMapper, Sys
 			return ApiResp.failure("待删除的权限模块还存在子权限模块, 无法删除");
 		}
 
+		Long aclCount = aclService.getAclCount(req.getSurrogateId());
+		if (aclCount >= 1) {
+			return ApiResp.failure("待删除的权限模块下还存在权限点, 无法删除");
+		}
+
 		aclModuleMapper.deleteById(aclModule.getId());
 		return ApiResp.success("删除权限模块成功");
+	}
+
+	/**
+	 * 获取单个权限模块对象
+	 * @param surrogateId
+	 * @return
+	 */
+	@Override
+	public ApiResp<SysAclModuleVO> getAclModule(Long surrogateId) {
+		QueryWrapper<SysAclModule> queryWrapper = new QueryWrapper<>();
+		queryWrapper.eq("surrogate_id", surrogateId);
+		SysAclModule aclModule = aclModuleMapper.selectOne(queryWrapper);
+		if (Objects.isNull(aclModule)) {
+			return ApiResp.failure(INFO_NOT_EXIST);
+		}
+		SysAclModuleVO aclModuleVO = new SysAclModuleVO();
+		BeanUtils.copyProperties(aclModule, aclModuleVO);
+		return ApiResp.success(aclModuleVO);
+	}
+
+	/**
+	 * 查询所有的权限模块列表
+	 * @return
+	 */
+	@Override
+	public ApiResp<List<SysAclModuleVO>> aclModuleList(AclModuleListReq req) {
+		List<SysAclModuleVO> sysAclModules = aclModuleMapper.selectAclModuleList(req);
+		return ApiResp.success(sysAclModules);
 	}
 }
